requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.34.4
geographiclib==1.52
geopy==2.2.0
gpxpy==1.5.0
idna==3.3
jsonpickle==2.2.0
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
scipy==1.8.1
six==1.16.0
urllib3==1.26.10
